# Slightly improve the performance of the HyperbolicGenerator