The past implementation generated random numbers and sorted them seperately.

I replaced this with an algorithm that directly generates sorted random numbers (Program 1 from this 1980 paper: [Generating Sorted Lists of Random Numbers](https://dl.acm.org/doi/10.1145/355900.355907)).

I tested the performance of this change using the following configuration:
```
	HyperbolicGenerator gen{20000000};
	auto graph = gen.generate();
```
Here are some numbers:
​ | before | after
--- | --- | ----
Total Runtime | 14.9 s | 12.9 s
Point generation runtime | 3.1 s | 1.3 s

Memory usage should also be a bit lower, because I removed the permutation and copy vectors.

Also note that the `fillPointsSorted` is basically a copy of the `fillPoints` functions. I kept the old `fillPoints` function because it is still being used by the DynamicHyperbolicGenerator and I didn't want to mess with that.