@IntegratedQuantum 2023-09-27
The CI failure was caused by an oversight in the test case.
The test case was modifying the graph size by calling
`extractLargestConnectedComponent`
But afterwards the test case was reusing the old size.
This went unnoticed because the particular seeds used to create fully connected graphs.

Anyways I fixed it now.