@IntegratedQuantum 2023-09-27
Now there is one CI failure that I don't understand.
The "Linux (gcc-11, CPython 3.11): documentation" CI is failing with the error:
```
Extension error (exhale):
Handler <function environment_ready at 0x7f562e01e3e0> for event 'builder-inited' threw an exception (exception: startswith first arg must be str or a tuple of str, not _StrPath)
```
I have no idea what this means or how to resolve it.