@fabratu 2023-10-06
You can (but don't have to) rebase the PR to resolve the CI failure. It was an error with offending package versions for the documentation build.