@IntegratedQuantum 2023-11-06
>  Did you benchmark against several parameters of HyperbolicGenerator?

I didn't think that's necessary, given that the code I touched only depends on the total number of nodes. So the absolute performance gain should be constant throughout different parameters.