@IntegratedQuantum 2023-11-09
I renamed the variable and as requested I also checked performance for some more values of n ranging from 10000 to 10000000, in all cases the new version appeared to be significantly faster.